Sample updated to Preview 6 changes
