--- src/TestApp/Android/MainActivity.cs
@@ -1,10 +1,0 @@
-﻿using Android.App;
-using Microsoft.Maui;
-
-namespace TestApp
-{
-	[Activity(Theme = "@style/AppTheme.NoActionBar", MainLauncher = true)]
-	public class MainActivity : MauiAppCompatActivity
-	{
-	}
-}
