--- src/TestApp/Android/Resources/values/colors.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="colorPrimary">#512BD4</color>
-    <color name="colorPrimaryDark">#DFD8F7</color>
-    <color name="colorAccent">#2B0B98</color>
-</resources>
