--- src/TestApp/Android/Resources/values/dimens.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<resources>
-    <dimen name="fab_margin">16dp</dimen>
-</resources>
