--- src/TestApp/Android/Resources/values/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<resources>
-    <string name="action_settings">Settings</string>
-</resources>
