--- src/TestApp/App.xaml
@@ -1,12 +1,0 @@
-﻿<Application xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:TestApp"
-             x:Class="TestApp.App">
-    <Application.Resources>
-        <ResourceDictionary>
-
-            <!-- Add any additional resources here. -->
-
-        </ResourceDictionary>
-    </Application.Resources>
-</Application>
