--- src/TestApp/App.xaml.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.Maui;
-using Microsoft.Maui.Controls;
-
-namespace TestApp
-{
-	public partial class App : Application
-	{
-		public App()
-		{
-			InitializeComponent();
-		}
-
-		public override IWindow CreateWindow(IActivationState activationState)
-		{
-			Microsoft.Maui.Controls.Compatibility.Forms.Init(activationState);
-
-			return new MainWindow();
-		}
-	}
-}
