--- src/TestApp/MainPage.xaml
@@ -1,29 +1,0 @@
-<core:MauiPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:core="clr-namespace:TestApp.Core"
-             x:Class="TestApp.MainPage"
-             BackgroundColor="White">
-
-    <StackLayout Padding="{OnPlatform iOS='30,60,30,30', Default='30'}">
-
-        <Label Text="Welcome to"
-               FontSize="32"
-               TextColor="Black"
-               HorizontalOptions="Center" />
-        <Label Text=".NET Multi-platform App UI"
-               FontSize="24"
-               TextColor="Black"
-               HorizontalOptions="Center" />
-
-        <Button Text="Click Me!"
-                HorizontalOptions="Center"
-                Clicked="OnButtonClicked"/>
-        <Label x:Name="CountLabel"
-               Text="Click the button above!"
-               FontSize="20"
-               TextColor="Black"
-               HorizontalOptions="Center" />
-
-    </StackLayout>
-
-</core:MauiPage>
