--- src/TestApp/MainWindow.cs
@@ -1,12 +1,0 @@
-using TestApp.Core;
-
-namespace TestApp
-{
-    public class MainWindow : Window
-	{
-		public MainWindow()
-		{
-			Page = new MainPage();
-		}
-	}
-}
