+++ src/TestApp/TestApp.WinUI/Properties/launchSettings.json
@@ -1,0 +1,7 @@
+{
+  "profiles": {
+    "TestApp.WinUI": {
+      "commandName": "MsixPackage"
+    }
+  }
+}
