+++ src/TestApp/TestApp.WinUI/TestApp.WinUI.csproj
@@ -1,0 +1,61 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFrameworks>net6.0-windows10.0.19041</TargetFrameworks>
+		<RuntimeIdentifier>win-x64</RuntimeIdentifier>
+		<OutputType>WinExe</OutputType>
+		<TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
+		<UseMaui>true</UseMaui>
+		<SingleProject>true</SingleProject>
+		<RootNamespace>TestApp</RootNamespace>
+
+		<!-- Required - Enable Single Project for WinUI -->
+		<EnablePreviewMsixTooling>true</EnablePreviewMsixTooling>
+		<WindowsPackageType>MSIX</WindowsPackageType>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<!-- App Icon -->
+		<MauiImage
+			Include="..\TestApp\Resources\appicon.svg"
+			Link="Resources\appicon.svg"
+			ForegroundFile="..\TestApp\Resources\appiconfg.svg"
+			IsAppIcon="true"
+			Color="#512BD4" />
+
+		<!-- Splash Screen -->
+		<MauiSplashScreen
+			Include="..\TestApp\Resources\appiconfg.svg"
+			Link="Resources\appiconfg.svg"
+			Color="#512BD4" />
+
+		<!-- Images -->
+		<MauiImage
+			Include="..\TestApp\Resources\Images\*"
+			Link="Resources\Images\%(Filename)%(Extension)" />
+
+		<!-- Custom Fonts -->
+		<MauiFont
+			Include="..\TestApp\Resources\Fonts\*"
+			Link="Resources\Fonts\%(Filename)%(Extension)" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<Compile
+			Include="..\TestApp\**\*.cs"
+			Exclude="..\TestApp\bin\**;..\TestApp\obj\**;..\TestApp\Platforms\**"
+			Link="%(RecursiveDir)%(Filename)%(Extension)"/>
+		<MauiXaml
+			Include="..\TestApp\**\*.xaml"
+			Exclude="..\TestApp\bin\**;..\TestApp\obj\**"
+			Link="%(RecursiveDir)%(FileName)%(Extension)" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.ProjectReunion" Version="0.8.0" />
+		<PackageReference Include="Microsoft.ProjectReunion.Foundation" Version="0.8.0" />
+		<PackageReference Include="Microsoft.ProjectReunion.WinUI" Version="0.8.0" />
+		<PackageReference Include="Microsoft.ProjectReunion.InteractiveExperiences" Version="0.8.0" NoWarn="NU1701" />
+	</ItemGroup>
+
+</Project>
