+++ src/TestApp/TestApp/MainPage.xaml.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using Microsoft.Maui.Controls;
+
+namespace TestApp
+{
+	public partial class MainPage : ContentPage
+	{
+		public MainPage()
+		{
+			InitializeComponent();
+		}
+
+		int count = 0;
+		private void OnCounterClicked(object sender, EventArgs e)
+		{
+			count++;
+			CounterLabel.Text = $"Current count: {count}";
+		}
+	}
+}
