+++ src/TestApp/TestApp/Platforms/Android/MainApplication.cs
