+++ src/TestApp/TestApp/Platforms/Android/Resources/values/styles.xml
