+++ src/TestApp/TestApp/Platforms/MacCatalyst/AppDelegate.cs
