+++ src/TestApp/TestApp/Platforms/MacCatalyst/Program.cs
