+++ src/TestApp/TestApp/Platforms/iOS/AppDelegate.cs
@@ -1,0 +1,10 @@
+﻿using Foundation;
+using Microsoft.Maui;
+
+namespace TestApp
+{
+    [Register(nameof(AppDelegate))]
+    public class AppDelegate : MauiUIApplicationDelegate<Startup>
+    {
+    }
+}
