+++ src/TestApp/TestApp/Startup.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.Maui;
+using Microsoft.Maui.Hosting;
+using Microsoft.Maui.Controls.Compatibility;
+using Microsoft.Maui.Controls.Hosting;
+using Microsoft.Maui.Controls.Xaml;
+
+[assembly: XamlCompilationAttribute(XamlCompilationOptions.Compile)]
+
+namespace TestApp
+{
+	public class Startup : IStartup
+	{
+		public void Configure(IAppHostBuilder appBuilder)
+		{
+			appBuilder
+				.UseMauiApp<App>()
+				.ConfigureFonts(fonts =>
+				{
+					fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
+				});
+		}
+	}
+}
