--- src/TestApp/global.json
@@ -1,7 +1,0 @@
-{
-  "sdk": {
-    "version": "6.0.100-preview.3.21202.5",
-    "allowPrerelease": true,
-    "rollForward": "disable"
-  }
-}
