--- src/TestApp/iOS/AppDelegate.cs
@@ -1,10 +1,0 @@
-﻿using Foundation;
-using Microsoft.Maui;
-
-namespace TestApp
-{
-	[Register("AppDelegate")]
-	public class AppDelegate : MauiUIApplicationDelegate<Startup>
-	{
-	}
-}
